third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/11.4.1-expected.txt

PASS Function("'use strict'; delete x;") threw exception of type SyntaxError.
PASS !!Function("delete x;") is true
PASS true === true
PASS Function("'use strict'; delete (x);") threw exception of type SyntaxError.
PASS !!Function("delete (x);") is true
PASS true === true
PASS !!Function("'use strict'; delete x.y;") is true
PASS !!Function("delete x.y;") is true
PASS true === true
PASS Function("'use strict'; function f() { delete x; }") threw exception of type SyntaxError.
PASS !!Function("function f() { delete x; }") is true
PASS true === true
PASS Function("'use strict'; function f() { \"use strict\"; delete x; }") threw exception of type SyntaxError.
PASS Function("function f() { \"use strict\"; delete x; }") threw exception of type SyntaxError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
